Backends/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.3.1
mysqlclient==2.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
